// src/app/models/year-end-inventory-models/sosCompletionReportModel.ts
/**
 * year-end-inventory-backend
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface SosCompletionReportModel { 
    plant?: number;
    salesOrder?: string;
    lineItem?: string;
    material?: string;
    counted?: number;
    sapCount?: number;
    difference?: number;
}