// src/app/models/year-end-inventory-models/uRCompletionReportModel.ts
/**
 * year-end-inventory-backend
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface URCompletionReportModel { 
    plant?: number;
    material?: string;
    storageLocation?: string;
    counted?: number;
    sapCount?: number;
    difference?: number;
}